+++ main.tf
@@ -76,6 +76,7 @@ module "account_service_role" {
   is_saas_mode                                = local.condition_is_saas_mode
   cloudscanner_saas_customer_assume_role_name = local.condition_create_customer_assume_role ? local.cloudscanner_saas_customer_assume_role_name : null
   permissions_boundary                        = var.account_service_role_permissions_boundary
+  upwind_enable_self_managed_kms_key          = var.upwind_enable_self_managed_kms_key
 }
 
 # Create the CloudScanner admin role. This will be in the orchestrator account.
@@ -399,7 +400,6 @@ module "cloudscanner_saas_customer_assume_role" {
   external_id                      = var.external_id
   cloudscanner_execution_role_name = local.cloudscanner_execution_role_name
   custom_tags                      = var.custom_tags
-  permissions_boundary             = var.cloudscanner_saas_customer_assume_role_permissions_boundary
 }
 
 # Using a null resource so that we can avail of the precondition rules to